fix: debug phases 3-6 - 5 bugs fixed, 6 regression tests added
Agent-Logs-Url: https://github.com/awest813/League-Soccer/sessions/43ee9836-049c-4369-8a6f-348049ff753e

+++ src/data/scoutingdata.cpp
@@ -35,7 +35,7 @@ ScoutReport ScoutingNetwork::GenerateReport(int scoutID, int playerID, int seaso
   int skill = scout ? scout->skill : 50;
 
   // Uncertainty is inversely proportional to scout skill: skill 100 = ±5, skill 0 = ±30
-  int baseUncertainty = 30 - (skill * 25 / 100);
+  int baseUncertainty = 30 - static_cast<int>(skill * 25.0f / 100.0f);
 
   static const char* attrs[] = {"pace", "shooting", "passing", "dribbling", "defending",
                                  "physical", nullptr};
